+++ src/CMakeLists.txt
@@ -164,7 +164,7 @@ add_library(utils SHARED
   utils/random.cpp
   utils/ScopedTimer.cpp
   utils/transform.cpp
-  utils/latch.cpp
+  utils/threading.cpp
   gps/gps_util.cpp
   base64/base64.cpp
   base64/base64_img.cpp
